+++ src/router.tsx
@@ -1,3 +1,4 @@
+import GeneratedData from "./components/User/Warning/GeneratedData";
 import { updatePageEvent } from "./firebase/config";
 import UserDeleteConfirmationPage from "./pages/admin/users/user-delete-confirmation";
 import { authClient } from "@/api/auth/client";
@@ -283,6 +284,10 @@ export default function Router (props: Props) {
                             <SuperAdminContentLibraryTable />
                         </Layout>
                     )}
+                />
+                <Route
+                    path="/user-deletion-warning"
+                    element={<GeneratedData />}
                 />
                 <Route
                     path="/user-deletion-confirmation"